requirements.txt
arrow==1.3.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
binaryornot==0.4.4
bs4==0.0.2
cachetools==6.2.1
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
comm==0.2.3
cookiecutter==2.6.0
cssselect==1.3.0
cssutils==2.11.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httplib2==0.31.0
idna==3.11
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.8.0
nest-asyncio==1.6.0
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.0
premailer==3.10.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pyxdg==0.28
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
selenium==4.36.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spyder-kernels==3.1.0
stack-data==0.6.3
text-unidecode==1.3
tornado==6.5.2
traitlets==5.14.3
trio==0.31.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20251008
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
websocket-client==1.9.0
wsproto==1.2.0
wurlitzer==3.1.1
